Added some extra verification for deck creation and deletion and added extra unit tests.

+++ MementoMori/MementoMori.Server/Service/DeckHelper.cs
@@ -85,6 +85,12 @@ namespace MementoMori.Server.Service
         public async Task<Guid> CreateDeckAsync (EditedDeckDTO createDeck, Guid requesterId)
         {
             Guid newDeckGuid = Guid.NewGuid();
+            if (createDeck.Deck.Title == "" || createDeck.Deck.Title.TrimStart(' ').Length == 0)
+            {
+                throw new ArgumentException();
+            }
+            var cards = createDeck.NewCards?.ToList() ?? [];
+            int cardCount = cards.Count; 
             Deck newDeck = new()
             {
                 Id = newDeckGuid,
@@ -96,18 +102,22 @@ namespace MementoMori.Server.Service
                 Rating = 0,
                 RatingCount = 0,
                 Modified = DateOnly.FromDateTime(DateTime.Now),
-                Cards = createDeck.NewCards?.ToList() ?? [],
-                CardCount = 0,
+                Cards = cards,
+                CardCount = cardCount,
             };
             _context.Decks.Add(newDeck);
             await _context.SaveChangesAsync();
             return newDeckGuid;
         }
-        public async Task DeleteDeckAsync(Guid deckId)
+        public async Task DeleteDeckAsync(Guid deckId, Guid requesterId)
         {
             var deck = _context.Decks.Include(d => d.Cards).FirstOrDefault(d => d.Id == deckId);
             if (deck != null)
             {
+                if (deck.CreatorId != requesterId)
+                {
+                    throw new UnauthorizedEditingException();
+                }
                 _context.Decks.Remove(deck);
                 await _context.SaveChangesAsync();
             }
